fix(VersionedGridFieldDetailForm): Save when publishing
The VersionedGridFieldDetailForm_ItemRequest#publish method was only calling $record->doPublish but not updating the record with the new data.  This makes it call $this->doSave first; this may be overkill and a bit of

    $form->saveInto($this->record);
    $this->record->write();

might be enough?

Also, changes the name of the VersionedGridFieldDetailForm_ItemRequest#publish method to VersionedGridFieldDetailForm_ItemRequest#doPublish, in keeping with the convention of doSave in GridFieldDetailForm_ItemRequest

Fixes Issue #19